src/inspection_service.py
import logging
import pandas as pd
from typing import Dict, Any, Tuple, List
# from google.cloud import dlp_v2 # DLPクライアント (別途インストール・設定が必要)
# from google.cloud import aiplatform # Vertex AIクライアント (別途インストール・設定が必要)
from . import config, schema_manager
from google.cloud import dlp_v2
from google.cloud.aiplatform.gapic import PredictionServiceClient
from google.api_core.client_options import ClientOptions
from google.protobuf import json_format
from google.protobuf.struct_pb2 import Value

# DLPクライアントの初期化 (コメントアウト)
# dlp_client = dlp_v2.DlpServiceClient()
# parent = f"projects/{config.GOOGLE_CLOUD_PROJECT}"

# LLMクライアント/モデルの初期化 (コメントアウト)
# aiplatform.init(project=config.GOOGLE_CLOUD_PROJECT, location='your-location') # 必要に応じてロケーション指定
# llm_model = aiplatform.Endpoint(endpoint_name=config.LLM_API_ENDPOINT) # 例: Vertex AI Endpoint
# または特定のモデルをロード

# ロガーの設定
logger = logging.getLogger(__name__)

class InspectionService:
    """データ検査を担当するサービスクラス"""

    def __init__(self):
        # DLPクライアントの初期化
        self.dlp_client = dlp_v2.DlpServiceClient()
        self.dlp_parent = f"projects/{config.GOOGLE_CLOUD_PROJECT}/locations/{config.DLP_API_LOCATION}"

        # Vertex AI Predictionクライアントの初期化
        client_options = ClientOptions(api_endpoint=config.LLM_API_ENDPOINT)
        self.prediction_client = PredictionServiceClient(client_options=client_options)
        self.llm_endpoint = f"projects/{config.GOOGLE_CLOUD_PROJECT}/locations/us-central1/endpoints/{config.LLM_MODEL_NAME}" # リージョンは環境に合わせて変更

    def inspect_data_dlp(self, df: pd.DataFrame, schema: dict) -> list:
        """DLP APIを使用してデータフレーム内の機密データを検査する (ダミー実装)"""
        logger.info("データ検査を開始 (ダミー実装)...")
        violations = []
        # ここに実際のDLP API呼び出しロジックを実装する
        # 例: dfを適切な形式に変換し、dlp_client.inspect_contentを呼び出す
        # ダミーとして、'email'列があれば違反とする
        if 'email' in df.columns:
            for index, row in df.iterrows():
                if pd.notna(row['email']): # NaNでないことを確認
                    violations.append({
                        "row_index": index,
                        "column_name": 'email',
                        "finding": "EMAIL_ADDRESS (dummy)",
                        "details": f"Found potentially sensitive data in row {index}, column 'email'."
                    })
        # logger.info(f"DLP検査完了 (ダミー): {len(violations)}件の違反候補")
        return violations

    def inspect_data_llm(self, df: pd.DataFrame, schema: dict) -> list:
        """LLM APIを使用してデータフレーム内の項目がスキーマ定義に準拠しているか検査する (ダミー実装)"""
        # logger.info("スキーマ準拠性検査を開始 (ダミー実装)...")
        violations = []
        # ここに実際のLLM API呼び出しロジックを実装する
        # 例: スキーマ定義とデータ行をプロンプトに含め、LLMに評価させる
        # ダミーとして、'age'列が数値でない、または負の値であれば違反とする
        if 'age' in df.columns and 'int' in schema.get('age', {}).get('type', '').lower():
            for index, row in df.iterrows():
                age = row['age']
                if pd.notna(age): # NaNでないことを確認
                    try:
                        age_val = int(age)
                        if age_val < 0:
                            violations.append({
                                "row_index": index,
                                "column_name": 'age',
                                "finding": "INVALID_VALUE (dummy)",
                                "details": f"Age cannot be negative in row {index}. Value: {age_val}"
                            })
                    except (ValueError, TypeError):
                        violations.append({
                            "row_index": index,
                            "column_name": 'age',
                            "finding": "INVALID_TYPE (dummy)",
                            "details": f"Age must be an integer in row {index}. Value: {age}"
                        })
        # logger.info(f"LLM検査完了 (ダミー): {len(violations)}件の違反候補")
        return violations

    def inspect_data(self, df: pd.DataFrame, schema: dict) -> list:
        """データフレームに対してDLPとLLMの両方の検査を実行する"""
        all_violations = []
        try:
            dlp_violations = self.inspect_data_dlp(df, schema)
            all_violations.extend(dlp_violations)
        except Exception as e:
            logger.error(f"DLP API呼び出し中にエラー: {e}", exc_info=True)
            # エラーが発生した場合でも、LLM検査は試みる
            all_violations.append({
                "row_index": -1, # 全体エラーを示す
                "column_name": "N/A",
                "finding": "DLP_API_ERROR",
                "details": f"Error during DLP inspection: {e}"
            })

        try:
            llm_violations = self.inspect_data_llm(df, schema)
            all_violations.extend(llm_violations)
        except Exception as e:
            logger.error(f"LLM API呼び出し中にエラー: {e}", exc_info=True)
            all_violations.append({
                "row_index": -1, # 全体エラーを示す
                "column_name": "N/A",
                "finding": "LLM_API_ERROR",
                "details": f"Error during LLM inspection: {e}"
            })

        if all_violations:
            logger.info(f"データ検査完了: {len(all_violations)} 件の違反を検出しました。")
        else:
            logger.info("データ検査完了: 違反はありませんでした。")

        return all_violations

    # --- 以下、実際のAPI呼び出しの参考例 (コメントアウト) ---
    # def _call_dlp_api(self, item_to_inspect):
    #     """実際のDLP APIを呼び出すメソッド (参考)"""
    #     try:
    #         response = self.dlp_client.inspect_content(
    #             request={
    #                 "parent": self.dlp_parent,
    #                 "inspect_config": {
    #                     "info_types": [{"name": "EMAIL_ADDRESS"}, {"name": "PHONE_NUMBER"}],
    #                     "min_likelihood": dlp_v2.Likelihood.POSSIBLE,
    #                     "include_quote": True,
    #                 },
    #                 "item": item_to_inspect, # dlp_v2.ContentItem 形式
    #             }
    #         )
    #         return response.result.findings
    #     except Exception as e:
    #         # print(f"DLP API呼び出し中にエラー: {e}")
    #         logger.error(f"DLP API呼び出し中にエラー: {e}", exc_info=True)
    #         return []
    #
    # def _call_llm_api(self, prompt):
    #     """実際のVertex AI Prediction APIを呼び出すメソッド (参考)"""
    #     try:
    #         instance = json_format.ParseDict({"prompt": prompt}, Value())
    #         instances = [instance]
    #         parameters_dict = {"temperature": 0.2, "maxOutputTokens": 256, "topP": 0.8, "topK": 40}
    #         parameters = json_format.ParseDict(parameters_dict, Value())
    #
    #         response = self.prediction_client.predict(
    #             endpoint=self.llm_endpoint,
    #             instances=instances,
    #             parameters=parameters,
    #         )
    #         # レスポンスの解析 (モデルによって異なる)
    #         # predictions = [json_format.MessageToDict(p) for p in response.predictions]
    #         return response.predictions # 仮
    #     except Exception as e:
    #         # print(f"LLM API呼び出し中にエラー: {e}")
    #         logger.error(f"LLM API呼び出し中にエラー: {e}", exc_info=True)
    #         return [] 
