pyproject.toml
[project]
name = "mock-mito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.87.0",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-dlp>=3.29.0",
    "mitosheet>=0.2.17",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "streamlit-option-menu>=0.3.9",
    "Flask>=3.0.3",
    "gunicorn>=22.0.0",
    "db-dtypes",
]
